-- fixed centos 5.5 build
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@130 01689507-d990-df11-9d7d-0017a411d416

+++ sources/common/include/platform/linux/linuxplatform.h
@@ -114,8 +114,8 @@ do { \
 #define GETCUSTOMNTP(result,value) \
 do { \
 	struct timeval tv; \
-	tv.tv_sec=value/CLOCKS_PER_SECOND; \
-	tv.tv_usec=value-tv.tv_sec*CLOCKS_PER_SECOND; \
+	tv.tv_sec=(uint64_t)value/CLOCKS_PER_SECOND; \
+	tv.tv_usec=(uint64_t)value-tv.tv_sec*CLOCKS_PER_SECOND; \
 	result=(((uint64_t)tv.tv_sec + 2208988800U)<<32)|((((uint32_t)tv.tv_usec) << 12) + (((uint32_t)tv.tv_usec) << 8) - ((((uint32_t)tv.tv_usec) * 1825) >> 5)); \
 }while (0);
